requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.5.2
httpx>=0.24.0,<0.25.0
python-jose==3.3.0
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.1.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
stripe==7.6.0
email-validator==2.1.0.post1
sse-starlette==1.6.5
